feat: separate system tests from normal tests
By introducing a feature flag `system_tests` we can run tests separately.
Make sure to always add the feature flag `--all-features` when running clippy or the tests locally
