+++ .github/workflows/ci.yml
@@ -50,7 +50,7 @@ jobs:
         run: cargo install flutter_rust_bridge_codegen
       - name: Generate FFI bindings
         run: just gen
-      - run: cargo clippy --all-targets -- -D warnings
+      - run: cargo clippy --all-targets --all-features -- -D warnings
 
   tests:
     runs-on: ubuntu-latest
