+++ crates/ln-dlc-node/Cargo.toml
@@ -35,3 +35,6 @@ tokio = { version = "1", default-features = false, features = ["io-util", "macro
 tracing = "0.1.37"
 tracing-log = "0.1.3"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+
+[features]
+system_tests = []
